Refactor display module to containerize curses
The curses interactions are mixed in with the creation and mapping of
colors, which makes it difficult to test. This change updates the
display module to provide a basic interface for ncurses and extracts any
non-trivial functionality into a curses agnostic file.
